Move GridView control tests into a separate file
