+++ src/Tests/ControlTests/SimpleControlTests.cs
@@ -212,80 +212,6 @@ namespace DotVVM.Framework.Tests.ControlTests
         }
 
         [TestMethod]
-        public async Task GridViewColumn_CellDecorators()
-        {
-            var r = await cth.RunPage(typeof(BasicTestViewModel), @"
-                <dot:GridView DataSource={value: Customers} RenderSettings.Mode=Server InlineEditing=true>
-                    <dot:GridViewTextColumn HeaderText=Id ValueBinding={value: Id}>
-                        <HeaderCellDecorators>
-                            <dot:Decorator class={value: 'header-' + Integer} />
-                        </HeaderCellDecorators>
-                        <CellDecorators>
-                            <dot:Decorator class={value: 'cell-' + Id} />
-                        </CellDecorators>
-                        <EditCellDecorators>
-                            <dot:Decorator class={value: 'cell-' + Id + '-edit'} />
-                        </EditCellDecorators>
-                    </dot:GridViewTextColumn>
-                </dot:GridView>");
-
-            Assert.IsTrue(r.Html.QuerySelectorAll("th")[0].ClassList.Contains("header-10000000"));
-            Assert.IsTrue(r.Html.QuerySelectorAll("td")[0].ClassList.Contains("cell-1-edit"));
-            Assert.IsTrue(r.Html.QuerySelectorAll("td")[1].ClassList.Contains("cell-2"));
-        }
-
-        [TestMethod]
-        public async Task GridViewColumn_HeaderRowDecorators()
-        {
-            var r = await cth.RunPage(typeof(BasicTestViewModel), @"
-                <dot:GridView DataSource={value: Customers} RenderSettings.Mode=Server InlineEditing=true>
-                    <HeaderRowDecorators>
-                        <dot:Decorator class={value: 'header-' + Integer} />
-                    </HeaderRowDecorators>
-                    <Columns>
-                        <dot:GridViewTextColumn HeaderText=Id ValueBinding={value: Id} />
-                    </Columns>
-                </dot:GridView>");
-
-            Assert.IsTrue(r.Html.QuerySelectorAll("tr")[0].ClassList.Contains("header-10000000"));
-        }
-
-        [TestMethod]
-        public async Task GridViewColumn_Usage_Validators()
-        {
-            var r = await cth.RunPage(typeof(BasicTestViewModel), @"
-                <dot:GridView DataSource={value: Customers} InlineEditing=true>
-                    <dot:GridViewTextColumn ValueBinding={value: Name} ValidatorPlacement=AttachToControl />
-                </dot:GridView>");
-
-            Assert.IsTrue(r.Html.QuerySelector("input[type=text]").GetAttribute("data-bind").Contains("dotvvm-validation: Name"));
-        }
-
-        [TestMethod]
-        public async Task GridViewColumn_Usage_AttachedProperties()
-        {
-            var exception = await Assert.ThrowsExceptionAsync<DotvvmCompilationException>(() => 
-                cth.RunPage(typeof(BasicTestViewModel), @"
-                <dot:GridView DataSource={value: Customers}>
-                    <dot:GridViewCheckBoxColumn Validation.Enabled=false ValueBinding={value: Enabled} />
-                </dot:GridView>"));
-
-            Assert.IsTrue(exception.Message.Contains("The column doesn't support the property Validation.Enabled! If you need to set an attached property applied to a table cell, use the CellDecorators property."));
-        }
-
-        [TestMethod]
-        public async Task GridViewColumn_Usage_DataContext()
-        {
-            var exception = await Assert.ThrowsExceptionAsync<DotvvmCompilationException>(() => 
-                cth.RunPage(typeof(BasicTestViewModel), @"
-                <dot:GridView DataSource={value: Customers}>
-                    <dot:GridViewTextColumn DataContext={value: _this} ValueBinding={value: Name} />
-                </dot:GridView>"));
-
-            Assert.IsTrue(exception.Message.Contains("Changing the DataContext property on the GridViewColumn is not supported!"));
-        }
-
-        [TestMethod]
         public async Task NamedCommand()
         {
             var r = await cth.RunPage(typeof(BasicTestViewModel), @"
